+++ PCP/src/test/java/PCP/Min/data/PCPMsgUserToUserPacket_Test.java
@@ -1,33 +1,35 @@
 /*
  * this is a school project under "The Unlicence".
  */
+package PCP.Min.data;
 
-package PCP.packets;
-
-import PCP.Min.data.MsgUserToGroup;
-import java.util.*;
+import PCP.Min.data.MsgUserToUser;
+import java.util.ArrayList;
 import org.junit.*;
 
 /**
  *
- * @author gfurri20
+ * @author JacopoWolf
  */
-public class PCPMsgUserToGroupPacket_Test
+public class PCPMsgUserToUserPacket_Test
 {
     @Test
     public void testToBytesSingle()
     {
-        MsgUserToGroup packet = new MsgUserToGroup
-        (
-            new byte[] {0,0},
-            "test"
+        MsgUserToUser packet = new MsgUserToUser
+        (   
+            new byte[]{0,0},
+            "testme",
+            "this is simple text"
         );
         
         byte[] expectedResult = new byte[]
         { 
-            05, 
+            01, 
             0,0,
-            116, 101, 115, 116,
+            116,101,115,116,109,101,
+            0,
+            116,104,105,115,32,105,115,32,115,105,109,112,108,101,32,116,101,120,116,
             0
         };
         
@@ -37,5 +39,7 @@ public class PCPMsgUserToGroupPacket_Test
             Assert.fail();
         
         Assert.assertArrayEquals( expectedResult, results.get(0) );
+        
     }
+    
 }
